requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
et-xmlfile==1.1.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
ttkthemes==3.2.2
tzdata==2024.2
